altera instancia st70.txt e transforma matriz de distancia para tipo (int). Tambem altera todas as variaveis fitness do tipo (double) para (int)

+++ src/individuo.cpp
@@ -20,7 +20,7 @@ void Individuo::setFitness(){
 
 }
 
-double Individuo::getFitness(){
+int Individuo::getFitness(){
 	return fitness;
 }
